crc16: move crc16 inside strkey/internal (#3932)
Move the `crc16` package to `strkey/internal/crc16`.

The `crc16` package implements a single form of crc16, the form specifically used by [strkeys](https://stellar.org/protocol/sep-23). It is only used by the `strkey` package and it is not really something that needs to be exported for others to use in other contexts.

The `crc16` package is not officially released in any particular package.
